+++ src/common/generate_wave.rs
@@ -13,7 +13,7 @@ use num_complex::Complex;
 /// * `num_samples` - The Number Of Total Samples To To Make
 ///
 /// * `offset` - The Number Of Samples To Skip (IE: You already made 600 samples and want the next 100)
-pub fn generate_wave(frequency: f32, sample_rate: f32, num_samples: i32, offset: i32, i_amplitude: f32, q_amplitude: f32, i_phase_offset: f32, q_phase_offset: f32) -> Vec<Complex<f32>> {
+pub fn generate_wave(frequency: f32, sample_rate: f32, num_samples: usize, offset: usize, i_amplitude: f32, q_amplitude: f32, i_phase_offset: f32, q_phase_offset: f32) -> Vec<Complex<f32>> {
     let mut arr: Vec<Complex<f32>> = Vec::with_capacity(num_samples as usize);
 
     // base
